pyproject.toml
[project]
name = "mypythonlearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "langchain-openai>=0.3.28",
    "mcp[cli]>=1.12.4",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
